64 bit stuff that I dont really understand
git-svn-id: https://wokjab.svn.sourceforge.net/svnroot/wokjab/wokjab/trunk/wokjab@241 a4058eb7-c820-0410-9b0b-94ec00e1a786

+++ plugins/P2PRandomNumber/session.cpp
@@ -327,12 +327,12 @@ Session::Message(WokXMLTag *tag)
 	}
 	if ( plain.size() == requiredusers.size() )
 	{
-		unsigned long int r = 0;
+		unsigned long long int r = 0;
 		
 		std::map <std::string, std::string>::iterator plainiter;
 		for ( plainiter = plain.begin() ; plainiter != plain.end() ; plainiter++)
 		{
-			unsigned long int x = strtoull(("0x" + plainiter->second).c_str(), NULL, 16);
+			unsigned long long int x = strtoull(("0x" + plainiter->second).c_str(), NULL, 16);
 			r ^= x;
 			
 		}
